orion-dynamic/requirements-test.txt
flask
pytest
mysql-connector-python
cryptography
pyjwt
ansible-runner
bcrypt
marshmallow
flask-limiter
docker
schedule
paramiko
pytest-cov
